+++ Sistema/Models/Pizza.cs
@@ -11,10 +11,10 @@ namespace Models
         public int PizzaID { get; set; }
         public string Nome { get; set; }
         public string Ingredientes { get; set; }
-        public double PrecoBroto { get; set; }
-        public double PrecoMedia { get; set; }
-        public double PrecoGrande { get; set; }
-        public double PrecoGigante { get; set; }
+        public Decimal PrecoBroto { get; set; }
+        public Decimal PrecoMedia { get; set; }
+        public Decimal PrecoGrande { get; set; }
+        public Decimal PrecoGigante { get; set; }
 
     }
 }
